+++ ContractManager/ContractBlueprint/Generate.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ContractManager.Mission;
+using System;
 using System.Collections.Generic;
 using System.Text;
 
@@ -7,7 +8,7 @@ namespace ContractManager.ContractBlueprint
     internal static class Generate
     {
         // unit-test method to create an example contract and write it to disk
-        internal static void Example001Contract()
+        internal static void ExampleContract001()
         {
             var contractToWrite = new ContractBlueprint
             {
@@ -88,7 +89,7 @@ namespace ContractManager.ContractBlueprint
             }
         }
 
-        internal static void Example002Contract()
+        internal static void ExampleContract002()
         {
             var contractToWrite = new ContractBlueprint
             {
@@ -195,6 +196,165 @@ namespace ContractManager.ContractBlueprint
             if (!string.IsNullOrEmpty(myDocumentsPath))
             {
                 contractToWrite.WriteToFile(savePath);
+            }
+        }
+
+        internal static void ExampleMission001()
+        {
+            // Mission
+            var missionToWrite = new MissionBlueprint
+            {
+                uid = "example_mission_001",
+                title = "Example Mission 001",
+                synopsis = "This is an example mission, with 2 contracts.",
+                description = "Orbit around Luna and fly back to Earth."
+            };
+
+            // Contract to Luna
+            var contractToLuna = new ContractBlueprint
+            {
+                uid = "example_mission_001_contract_001",
+                title = "Fly to Luna",
+                synopsis = "Fly to Luna and achieve a stable orbit",
+                description = "Create a randevous trajectory with Luna. Then achieve an orbit around Luna below 30km.",
+                isAutoAccepted = true,
+                isRejectable = false,
+            };
+            contractToLuna.prerequisites.Add(new Prerequisite
+            {
+                type = PrerequisiteType.MaxNumOfferedContracts,
+                maxNumOfferedContracts = 4,
+            });
+            // Not needed to check if the mission has been accepted, this is already built in.
+            
+            contractToLuna.requirements.Add(new Requirement
+            {
+                uid = "orbit_earth",
+                type = RequirementType.Orbit,
+                title = "Orbit Earth",
+                synopsis = "Starting condition is to orbit Earth below 210km.",
+                orbit = new RequiredOrbit
+                {
+                    targetBody = "Earth",
+                    type = OrbitType.Elliptical,
+                    maxApoapsis = 210000,
+                }
+            });
+            contractToLuna.requirements.Add(new Requirement
+            {
+                uid = "to_luna",
+                type = RequirementType.Orbit,
+                title = "Rendevous with Luna",
+                synopsis = "Rendevous with Luna with periapsis below 200km.",
+                description = "Change your orbit around Earth such that your trajectory will rendevous with Luna. TODO: Add hints.",
+                orbit = new RequiredOrbit
+                {
+                    targetBody = "Luna",
+                    maxPeriapsis = 200000,
+                    type = OrbitType.Escape,
+                }
+            });
+            contractToLuna.requirements.Add(new Requirement
+            {
+                uid = "orbit_luna",
+                type = RequirementType.Orbit,
+                title = "Orbit Luna",
+                synopsis = "Orbit Luna with apoapsis below 30km.",
+                description = "Change your orbit around Luna to a low orbit of below 30km.",
+                orbit = new RequiredOrbit
+                {
+                    targetBody = "Luna",
+                    maxApoapsis = 30000,
+                    minApoapsis = 10000,
+                    maxPeriapsis = 30000,
+                    minPeriapsis = 10000,
+                    type = OrbitType.Elliptical,
+                }
+            });
+            
+            // Contract back to Earth
+            var contractToEarth = new ContractBlueprint
+            {
+                uid = "example_mission_001_contract_002",
+                title = "Fly to Luna",
+                synopsis = "Fly to Luna and achieve a stable orbit",
+                description = "Create an escape trajectory from Luna. Then achieve an orbit around Earth between 150~200km.",
+                isAutoAccepted = true,
+                isRejectable = false,
+            };
+            contractToEarth.prerequisites.Add(new Prerequisite
+            {
+                type = PrerequisiteType.HasCompletedContract,
+                hasCompletedContract = "example_mission_001_contract_001",
+            });
+            contractToEarth.requirements.Add(new Requirement
+            {
+                uid = "escape_luna",
+                type = RequirementType.Orbit,
+                title = "Escape Luna",
+                synopsis = "Escape Luna back to Earth.",
+                description = "Change your orbit around Luna to a low orbit of below 30km.",
+                orbit = new RequiredOrbit
+                {
+                    targetBody = "Luna",
+                    type = OrbitType.Escape,
+                }
+            });
+            contractToEarth.requirements.Add(new Requirement
+            {
+                uid = "orbit_earth_again",
+                type = RequirementType.Orbit,
+                title = "Orbit Earth again",
+                synopsis = "Orbit Earth again.",
+                description = "Change your orbit around Earth to a orbit of between 150~200km.",
+                orbit = new RequiredOrbit
+                {
+                    targetBody = "Earth",
+                    type = OrbitType.Elliptical,
+                    maxApoapsis = 200000,
+                    maxPeriapsis = 200000,
+                    minApoapsis = 150000,
+                    minPeriapsis = 150000,
+                }
+            });
+            missionToWrite.contractBlueprintUIDs.Add(contractToLuna.uid);
+            missionToWrite.contractBlueprintUIDs.Add(contractToEarth.uid);
+            contractToLuna.missionBlueprintUID = missionToWrite.uid;
+            contractToEarth.missionBlueprintUID = missionToWrite.uid;
+            
+            string myDocumentsPath = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+            Console.WriteLine($"[CM] 'My Documents' path: {myDocumentsPath}");
+            string savePath = Path.Combine(
+                myDocumentsPath,
+                @"My Games\Kitten Space Agency\contracts\",
+                $"{contractToLuna.uid}.xml"
+            );
+            Console.WriteLine($"[CM] save path: {savePath}");
+            if (!string.IsNullOrEmpty(myDocumentsPath))
+            {
+                contractToLuna.WriteToFile(savePath);
+            }
+            
+            savePath = Path.Combine(
+                myDocumentsPath,
+                @"My Games\Kitten Space Agency\contracts\",
+                $"{contractToEarth.uid}.xml"
+            );
+            Console.WriteLine($"[CM] save path: {savePath}");
+            if (!string.IsNullOrEmpty(myDocumentsPath))
+            {
+                contractToEarth.WriteToFile(savePath);
+            }
+            
+            savePath = Path.Combine(
+                myDocumentsPath,
+                @"My Games\Kitten Space Agency\missions\",
+                $"{missionToWrite.uid}.xml"
+            );
+            Console.WriteLine($"[CM] save path: {savePath}");
+            if (!string.IsNullOrEmpty(myDocumentsPath))
+            {
+                missionToWrite.WriteToFile(savePath);
             }
         }
     }